Flatten attribute groups
Notes:
Similar to the global simpe types, copy attributes, type, values to
referencing elements from attribute groups. The goal is to generate
as little classes as possible even if we have to repeat fields.

That's a nicer strategy than having multi-inheritance in the final
data classes and paves the way for fields ordering :)