Remove the internal prefix for value fields
Notes:
In the past enumeration classes were allowed
to be inner|nested. To resolve that issues
I had used the prefix @ for value fields
in order to quickly identify these cases.

I know it was an ugly hack but worked well
till now that someone noticed the conflict
with the originalCase name convention.

Since that part of the code was hackish
and not really used anymore we can get rid
for the code and the silly prefix.