This is a prototype for #854
When a restriction is set, the ideal situation for the datamodel would be that only the properties that fall under the restriction are available.

From https://www.reddit.com/r/learnpython/comments/12qhzi6/dataclasses_with_inheritance/ I found that kw_only was used to resolve the issue with non-default argument 'class name' follows default argument.